requirements.txt
django==5.2
djangorestframework==3.16
djangorestframework-simplejwt==5.5
django-allauth==65.9
dj-rest-auth==7.0
django-environ
mysqlclient==2.2.2
django-mysql==4.17